+++ src/objects.cc
@@ -111,12 +111,12 @@ Handle<Map> Map::Rehash( GC* gc , const Handle<Map>& old_map ) {
 /* ---------------------------------------------------------------
  * Prototype
  * --------------------------------------------------------------*/
-Prototype::Prototype( const Handle<String>& pp , std::size_t argument_size ,
-                                                 std::size_t max_local_var_size,
-                                                 std::size_t real_table_size,
-                                                 std::size_t string_table_size,
-                                                 std::size_t upvalue_size,
-                                                 std::size_t code_buffer_size ,
+Prototype::Prototype( const Handle<String>& pp , std::uint8_t argument_size ,
+                                                 std::uint8_t max_local_var_size,
+                                                 std::uint8_t real_table_size,
+                                                 std::uint8_t string_table_size,
+                                                 std::uint8_t upvalue_size,
+                                                 std::uint32_t code_buffer_size ,
                                                  double* rtable,
                                                  String*** stable,
                                                  std::uint32_t* utable,
